Removed function base
