+++ nevergrad/benchmark/execution.py
@@ -5,7 +5,7 @@
 import heapq
 from collections import deque
 from typing import List, Callable, Any, NamedTuple, Tuple, Dict, Optional, Deque
-from ..functions.base import PostponedObject  # this object only serves to provide delays that the executor must use to order jobs
+from ..functions import PostponedObject  # this object only serves to provide delays that the executor must use to order jobs
 
 
 class MockedTimedJob:
