+++ nevergrad/functions/utils.py
@@ -3,7 +3,8 @@
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
 
-from typing import List, Optional
+import abc
+from typing import Dict, Any, Tuple, List, Optional
 import numpy as np
 
 
@@ -26,3 +27,18 @@ class Transform:
         if self.rotation_matrix is not None:
             y = self.rotation_matrix.dot(y)
         return y
+
+
+class PostponedObject(abc.ABC):
+    """Abstract class to inherit in order to notify the steady state benchmark executor that
+    the function implements a delay. This delay will be used while benchmarking to provide the
+    evaluation in a varying order.
+    The main aim of this class is to make sure there is no typo in the name of the special function.
+
+    See benchmark/execution.py for more details. This object is implemented here to avoid circular
+    imports.
+    """
+
+    @abc.abstractmethod
+    def get_postponing_delay(self, arguments: Tuple[Tuple[Any, ...], Dict[str, Any]], value: float) -> float:
+        raise NotImplementedError
